Working/requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
pandas>=2.0.0
seaborn>=0.12.0
icrawler>=0.6.0
kaggle>=1.5.0

